Create better code structure with authService
