+++ frontend/src/Components/Login.jsx
@@ -1,8 +1,7 @@
 import { useState } from "react";
 import styled from "styled-components";
 import { Link, useNavigate } from "react-router-dom";
-import { getAuth, signInWithEmailAndPassword } from "firebase/auth";
-import firebaseApp from "../config/authconfig";
+import authService from "../services/authService";
 
 const Container = styled.div`
   position: relative;
@@ -110,26 +109,9 @@ const LoginForm = () => {
   const handleSignIn = async (e) => {
     e.preventDefault();
     const { email, password } = formData;
-    console.log(email);
 
-    const auth = getAuth(firebaseApp);
-    console.log(auth);
-    try {
-      const userCredential = await signInWithEmailAndPassword(
-        auth,
-        email,
-        password
-      );
-      console.log(userCredential);
-      const user = userCredential.user;
-      const idToken = await user.getIdToken();
-      localStorage.setItem("token", idToken);
-      alert("Login successful!");
+    if (await authService.isValidLogin(email, password)) {
       navigate("/");
-    } catch (error) {
-      console.log(error);
-      alert("Login failed. Check your credentials.");
-      console.error(error);
     }
   };
