+++ frontend/src/services/openAIService.js
@@ -1,10 +1,12 @@
 import axios from "axios";
 
+// Backend is opened on lohalhost:3000 and is taking care of prompting the model.
 const BASE_URL = "http://localhost:3000";
 
-const responseService = {
-  dummy: async () => await axios.get(`${BASE_URL}/`),
+const openAIService = {
+  dummy: async () => await axios.get(`${BASE_URL}/`), // Dummy function to test connection
   getExercise: async (conf) => {
+    // Function to get exercise from backend
     console.log("conf", conf);
     try {
       const response = await axios.post(`${BASE_URL}/generate-exercise`, conf);
@@ -16,4 +18,4 @@ const responseService = {
   },
 };
 
-export default responseService;
+export default openAIService;
